Add "eigs" function for generalized eigenvalue problems
This commit introduces an "eigs" function similar to the one in Matlab to find a specified number of eigenpairs from a generalized problem. It works by implementing the shift-invert method on top of the existing partial_eigen function. The implementation is generic but relies on a trait that is initially only implemented for SparseColMatRef.